Add files via upload
The main functions you want to look at are "getWeather" and "animateWeather" . The other functions assume you are using a micro controller called the "Teensy" and have extensive knowledge on the circuit design requirements. This code was used for a finished project called "Data Central" which is located in Albuquerque New Mexico, created by the Social Media Work Group.